+++ src/utils/index.ts
@@ -3,14 +3,13 @@ export * from './asyncUtils';
 export { ObserverClass } from './classObserver';
 export type { ObserverIdentifier, ObserverInterface } from './classObserver';
 export * as Color from './color';
+export * from './dateHelpers';
 export * from './deepLinkParams';
 export * from './filterUtils';
 export * from './formatString';
-export * from './formatTimeHelpers';
 export * from './getKeyByValue';
 export * from './jsonUtils';
 export * from './logger';
 export type { LoggerConfiguration, SubscriberCallbackFunction } from './logger';
 export * from './objectUtils';
 export * from './randomIdGenerator';
-export * from './screenUtils';
